// apps/web/src/lib/api/generated/model/searchVisitsStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews
 * OpenAPI spec version: 1.0.0
 */

export type SearchVisitsStatus =
  (typeof SearchVisitsStatus)[keyof typeof SearchVisitsStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SearchVisitsStatus = {
  SCHEDULED: 'SCHEDULED',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;