// apps/web/src/lib/api/generated/model/searchVisitsParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews
 * OpenAPI spec version: 1.0.0
 */
import type { SearchVisitsStatus } from './searchVisitsStatus';
import type { Pageable } from './pageable';

export type SearchVisitsParams = {
  vetId?: number;
  petId?: number;
  status?: SearchVisitsStatus;
  from?: string;
  to?: string;
  pageable: Pageable;
};